refactor(relation): split relation-graph into sub-modules (#595)
Decompose the 1235-line RelationGraph god node into cache, cycle detection,
query, traversal, and mutation modules while preserving the public API and
import path for recall ranking and anchor map consumers.
