+++ packages/memento-core/src/domains/relation/services/relation-graph-cycle-detector.ts
@@ -1,0 +1,103 @@
+/**
+ * 관계 그래프 순환 참조 감지 (DFS)
+ */
+
+import Database from 'better-sqlite3';
+import { LIMITS } from '../../../shared/constants/relation-constants.js';
+import type { RelationType } from '../../../shared/types/relation.js';
+import { DatabaseUtils } from '../../../shared/utils/database.js';
+import { logger } from '../../../shared/utils/logger.js';
+
+export class RelationGraphCycleDetector {
+  constructor(private db: Database.Database) {}
+
+  /**
+   * 순환 참조 감지 내부 로직 (트랜잭션 없이 실행)
+   */
+  async detectCycleInternal(
+    sourceId: string,
+    targetId: string,
+    relationType: RelationType,
+    maxDepth: number = LIMITS.MAX_CYCLE_DEPTH
+  ): Promise<boolean> {
+    if (sourceId === targetId) {
+      return false;
+    }
+
+    const visited = new Set<string>();
+    const nodeRelations = new Map<string, string[]>();
+
+    const dfs = async (currentId: string, target: string, depth: number): Promise<boolean> => {
+      if (depth > maxDepth) {
+        logger.warn('순환 참조 감지: 최대 탐색 깊이 초과', {
+          sourceId,
+          targetId,
+          currentId,
+          depth,
+          maxDepth
+        });
+        return false;
+      }
+
+      if (currentId === target) {
+        return true;
+      }
+
+      if (visited.has(currentId)) {
+        return false;
+      }
+
+      visited.add(currentId);
+
+      let targetIds: string[] = [];
+      if (nodeRelations.has(currentId)) {
+        targetIds = nodeRelations.get(currentId)!;
+      } else {
+        const rows = DatabaseUtils.all(this.db, `
+          SELECT target_id
+          FROM memory_relation
+          WHERE source_id = ? AND relation_type = ?
+        `, [currentId, relationType]);
+
+        for (const row of rows) {
+          if (typeof row === 'object' && row !== null && 'target_id' in row) {
+            const targetIdValue = (row as { target_id: unknown }).target_id;
+            if (typeof targetIdValue === 'string') {
+              targetIds.push(targetIdValue);
+            }
+          }
+        }
+
+        nodeRelations.set(currentId, targetIds);
+      }
+
+      for (const nextId of targetIds) {
+        if (await dfs(nextId, target, depth + 1)) {
+          return true;
+        }
+      }
+
+      return false;
+    };
+
+    return await dfs(targetId, sourceId, 0);
+  }
+
+  /**
+   * 순환 참조 감지 (공개 메서드, 트랜잭션 자동 관리)
+   */
+  async detectCycle(
+    sourceId: string,
+    targetId: string,
+    relationType: RelationType,
+    maxDepth: number = LIMITS.MAX_CYCLE_DEPTH
+  ): Promise<boolean> {
+    if (DatabaseUtils.isInTransaction(this.db)) {
+      return await this.detectCycleInternal(sourceId, targetId, relationType, maxDepth);
+    }
+
+    return await DatabaseUtils.runTransaction(this.db, async () => {
+      return await this.detectCycleInternal(sourceId, targetId, relationType, maxDepth);
+    });
+  }
+}
